pip-requirements.txt
requests==2.22.0
pytz==2018.9
mock==3.0.5
freezegun==0.3.11
pytest==4.6.5
marshmallow==2.19.5
attrs==19.3.0
prometheus_client==0.8.0
pathspec==0.8.0
futures==3.3.0
